docs(tutorial/ts): delete tutorial sample in favor of toh-6
