--- public/docs/_examples/tutorial/ts/app/hero.ts
@@ -1,4 +1,0 @@
-export class Hero {
-	id: number;
-	name: string;
-}
