--- public/docs/_examples/tutorial/ts/app/main.ts
@@ -1,10 +1,0 @@
-import { bootstrap } from '@angular/platform-browser-dynamic';
-import { ROUTER_PROVIDERS } from '@angular/router-deprecated';
-
-import { HeroService } from './hero.service';
-import { AppComponent } from './app.component';
-
-bootstrap(AppComponent, [
-  ROUTER_PROVIDERS,
-  HeroService
-]);
