--- public/docs/_examples/tutorial/ts/plnkr.json
@@ -1,8 +1,0 @@
-{
-  "description": "Tour of Heroes",
-  "files":[
-    "!**/*.d.ts",
-    "!**/*.js"
-  ],
-  "tags": ["tutorial", "tour", "heroes"]
-}
