+++ public/docs/ts/latest/tutorial/index.jade
@@ -27,7 +27,7 @@ include ../_util-fns
 // #enddocregion intro
 
 :marked
-  [Run the live example](/resources/live-examples/tutorial/ts/plnkr.html)
+  [Run the live example](/resources/live-examples/toh-6/ts/plnkr.html)
 
 // #docregion main
 .l-main-section
